+++ src/Twilio/Rest/Client.php
@@ -31,6 +31,7 @@ use Twilio\Base\BaseClient as BaseClient;
  * @property Intelligence $intelligence
  * @property IpMessaging $ipMessaging
  * @property Lookups $lookups
+ * @property Marketplace $marketplace
  * @property Messaging $messaging
  * @property Microvisor $microvisor
  * @property Monitor $monitor
@@ -109,6 +110,7 @@ class Client extends BaseClient {
     protected $_intelligence;
     protected $_ipMessaging;
     protected $_lookups;
+    protected $_marketplace;
     protected $_messaging;
     protected $_microvisor;
     protected $_monitor;
@@ -273,6 +275,17 @@ class Client extends BaseClient {
             $this->_lookups = new Lookups($this);
         }
         return $this->_lookups;
+    }
+    /**
+     * Access the Marketplace Twilio Domain
+     *
+     * @return Marketplace Marketplace Twilio Domain
+     */
+    protected function getMarketplace(): Marketplace {
+        if (!$this->_marketplace) {
+            $this->_marketplace = new Marketplace($this);
+        }
+        return $this->_marketplace;
     }
     /**
      * Access the Messaging Twilio Domain