+++ src/Twilio/Rest/Content/V2/ContentAndApprovalsList.php
@@ -17,9 +17,11 @@
 namespace Twilio\Rest\Content\V2;
 
 use Twilio\ListResource;
+use Twilio\Options;
 use Twilio\Stream;
 use Twilio\Values;
 use Twilio\Version;
+use Twilio\Serialize;
 
 
 class ContentAndApprovalsList extends ListResource
@@ -46,6 +48,7 @@ class ContentAndApprovalsList extends ListResource
      * Unlike stream(), this operation is eager and will load `limit` records into
      * memory before returning.
      *
+     * @param array|Options $options Optional Arguments
      * @param int $limit Upper limit for the number of records to return. read()
      *                   guarantees to never return more than limit.  Default is no
      *                   limit
@@ -56,9 +59,9 @@ class ContentAndApprovalsList extends ListResource
      *                        efficient page size, i.e. min(limit, 1000)
      * @return ContentAndApprovalsInstance[] Array of results
      */
-    public function read(int $limit = null, $pageSize = null): array
+    public function read(array $options = [], int $limit = null, $pageSize = null): array
     {
-        return \iterator_to_array($this->stream($limit, $pageSize), false);
+        return \iterator_to_array($this->stream($options, $limit, $pageSize), false);
     }
 
     /**
@@ -69,6 +72,7 @@ class ContentAndApprovalsList extends ListResource
      * The results are returned as a generator, so this operation is memory
      * efficient.
      *
+     * @param array|Options $options Optional Arguments
      * @param int $limit Upper limit for the number of records to return. stream()
      *                   guarantees to never return more than limit.  Default is no
      *                   limit
@@ -79,11 +83,11 @@ class ContentAndApprovalsList extends ListResource
      *                        efficient page size, i.e. min(limit, 1000)
      * @return Stream stream of results
      */
-    public function stream(int $limit = null, $pageSize = null): Stream
+    public function stream(array $options = [], int $limit = null, $pageSize = null): Stream
     {
         $limits = $this->version->readLimits($limit, $pageSize);
 
-        $page = $this->page($limits['pageSize']);
+        $page = $this->page($options, $limits['pageSize']);
 
         return $this->version->stream($page, $limits['limit'], $limits['pageLimit']);
     }
@@ -98,13 +102,33 @@ class ContentAndApprovalsList extends ListResource
      * @return ContentAndApprovalsPage Page of ContentAndApprovalsInstance
      */
     public function page(
+        array $options = [],
         $pageSize = Values::NONE,
         string $pageToken = Values::NONE,
         $pageNumber = Values::NONE
     ): ContentAndApprovalsPage
     {
+        $options = new Values($options);
 
         $params = Values::of([
+            'SortByDate' =>
+                $options['sortByDate'],
+            'SortByContentName' =>
+                $options['sortByContentName'],
+            'DateCreatedAfter' =>
+                Serialize::iso8601DateTime($options['dateCreatedAfter']),
+            'DateCreatedBefore' =>
+                Serialize::iso8601DateTime($options['dateCreatedBefore']),
+            'ContentName' =>
+                $options['contentName'],
+            'Content' =>
+                $options['content'],
+            'Language' =>
+                Serialize::map($options['language'], function ($e) { return $e; }),
+            'ContentType' =>
+                Serialize::map($options['contentType'], function ($e) { return $e; }),
+            'ChannelEligibility' =>
+                Serialize::map($options['channelEligibility'], function ($e) { return $e; }),
             'PageToken' => $pageToken,
             'Page' => $pageNumber,
             'PageSize' => $pageSize,