+++ src/Twilio/Rest/Marketplace.php
@@ -1,0 +1,5 @@
+<?php
+namespace Twilio\Rest;
+
+class Marketplace extends MarketplaceBase {
+}