+++ src/Twilio/Rest/Numbers/V1.php
@@ -26,6 +26,7 @@ use Twilio\Rest\Numbers\V1\PortingPortabilityList;
 use Twilio\Rest\Numbers\V1\PortingWebhookConfigurationList;
 use Twilio\Rest\Numbers\V1\PortingWebhookConfigurationDeleteList;
 use Twilio\Rest\Numbers\V1\PortingWebhookConfigurationFetchList;
+use Twilio\Rest\Numbers\V1\SigningRequestConfigurationList;
 use Twilio\Version;
 
 /**
@@ -37,6 +38,7 @@ use Twilio\Version;
  * @property PortingWebhookConfigurationList $portingWebhookConfigurations
  * @property PortingWebhookConfigurationDeleteList $portingWebhookConfigurationsDelete
  * @property PortingWebhookConfigurationFetchList $portingWebhookConfigurationFetch
+ * @property SigningRequestConfigurationList $signingRequestConfigurations
  * @method \Twilio\Rest\Numbers\V1\BulkEligibilityContext bulkEligibilities(string $requestId)
  * @method \Twilio\Rest\Numbers\V1\PortingPortInContext portingPortIns(string $portInRequestSid)
  * @method \Twilio\Rest\Numbers\V1\PortingPortInPhoneNumberContext portingPortInPhoneNumber(string $portInRequestSid, string $phoneNumberSid)
@@ -53,6 +55,7 @@ class V1 extends Version
     protected $_portingWebhookConfigurations;
     protected $_portingWebhookConfigurationsDelete;
     protected $_portingWebhookConfigurationFetch;
+    protected $_signingRequestConfigurations;
 
     /**
      * Construct the V1 version of Numbers
@@ -127,6 +130,14 @@ class V1 extends Version
             $this->_portingWebhookConfigurationFetch = new PortingWebhookConfigurationFetchList($this);
         }
         return $this->_portingWebhookConfigurationFetch;
+    }
+
+    protected function getSigningRequestConfigurations(): SigningRequestConfigurationList
+    {
+        if (!$this->_signingRequestConfigurations) {
+            $this->_signingRequestConfigurations = new SigningRequestConfigurationList($this);
+        }
+        return $this->_signingRequestConfigurations;
     }
 
     /**