+++ CHANGES.md
@@ -1,6 +1,18 @@
 twilio-php Changelog
 ====================
 
+[2024-06-06] Version 8.1.1
+--------------------------
+**Api**
+- Mark MaxPrice as obsolete
+
+**Lookups**
+- Update examples for `phone_number_quality_score`
+
+**Messaging**
+- List tollfree verifications on parent account and all sub-accounts
+
+
 [2024-05-24] Version 8.1.0
 --------------------------
 **Library - Docs**